requirements.txt
# ezMIgrations Requirements

# Runtime dependencies
PyYAML>=6.0  # For configuration file support

# Testing dependencies
pytest>=7.0.0




